Add backend for github badges feature
